+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/combat/KillAura.kt
@@ -81,9 +81,8 @@ object KillAura : Module("KillAura", Category.COMBAT, Keyboard.KEY_R) {
     private val simulateCooldown by boolean("SimulateCooldown", false)
     private val simulateDoubleClicking by boolean("SimulateDoubleClicking", false) { !simulateCooldown }
 
-    // CPS / Attack speed
     private val cps by intRange("CPS", 5..8, 1..50) { !simulateCooldown }.onChanged {
-        attackDelay = randomClickDelay(it.first, it.last)
+        attackDelay = randomClickDelay(it)
     }
 
     private val hurtTime by int("HurtTime", 10, 0..10) { !simulateCooldown && !SmartHit.handleEvents() }